@@ -16,6 +16,8 @@ import {
   Question,
   DocumentAutoFlow,
   MoveDocument,
+  StationWidgetData,
+  DocumentGenerationStatus,
 } from 'src/models';
 import { environment } from 'src/environments/environment';
 
@@ -359,26 +361,26 @@ export class DocumentService {
   /**
    * Creates a new document.
    *
-   * @param stationRithmId The station where we want to create a new document.
-   * @returns The id of the document object that has been saved.
+   * @param name The name of document.
+   * @param priority The priority of the document.
+   * @param stationRithmId The rithmid of the station where this document should start.
+   * @returns Return string rithmId.
    */
-  createNewDocument(stationRithmId: string): Observable<string> {
-    const documentResponse = {
-      rithmId: '78DF8E53-549E-44CD-8056-A2CBA055F32F',
-      name: '',
-      priority: 0,
-      currentStations: [
-        {
-          name: 'So long',
-          instructions: '',
-          rithmId: stationRithmId,
-          assignedUser: null,
-        },
-      ],
-      children: [],
-      parents: [],
+  createNewDocument(
+    name: string,
+    priority: number,
+    stationRithmId: string
+  ): Observable<string> {
+    const requestObject = {
+      name,
+      priority,
     };
-    return of(documentResponse.rithmId).pipe(delay(1000));
+    return this.http
+      .post<{ /** Document Rithm Id. */ rithmId: string }>(
+        `${environment.baseApiUrl}${MICROSERVICE_PATH}?stationRithmId=${stationRithmId}`,
+        requestObject
+      )
+      .pipe(map((response) => response.rithmId));
   }
 
   /**
@@ -403,5 +405,52 @@ export class DocumentService {
       `${environment.baseApiUrl}${MICROSERVICE_PATH}/assign-user`,
       requestObject
     );
+  }
+
+  /**
+   * Get document for station widgets.
+   *
+   * @param stationRithmId The Specific ID of station.
+   * @returns Returns data station widget.
+   */
+  getStationWidgetDocuments(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    stationRithmId: string
+  ): Observable<StationWidgetData> {
+    const dataWidgetStation: StationWidgetData = {
+      stationName: 'Dev1',
+      documentGeneratorStatus: DocumentGenerationStatus.Manual,
+      documents: [
+        {
+          rithmId: '123-123-123',
+          name: 'Granola',
+          priority: 1,
+          flowedTimeUTC: '2022-01-13T16:43:57.901Z',
+          lastUpdatedUTC: '2022-01-13T16:43:57.901Z',
+          assignedUser: {
+            rithmId: '123-123-123',
+            firstName: 'Pedro',
+            lastName: 'Jeria',
+            email: 'pablo@mundo.com',
+            isAssigned: true,
+          },
+        },
+        {
+          rithmId: '321-123-123',
+          name: 'Almond',
+          priority: 3,
+          flowedTimeUTC: '2022-01-15T16:43:57.901Z',
+          lastUpdatedUTC: '2022-01-15T16:43:57.901Z',
+          assignedUser: {
+            rithmId: '321-123-123',
+            firstName: 'Pablo',
+            lastName: 'Santos',
+            email: 'Jaime@mundo2.com',
+            isAssigned: true,
+          },
+        },
+      ],
+    };
+    return of(dataWidgetStation).pipe(delay(1000));
   }
 }
