@@ -6,11 +6,20 @@ import { StationCardComponent } from './station-card/station-card.component';
 import { DocumentListCardComponent } from './document-list-card/document-list-card.component';
 import { HeaderComponent } from './header/header.component';
 import { MyStationsComponent } from './my-stations/my-stations.component';
-import { SharedModule } from '../shared/shared.module';
 import { PreviouslyStartedDocumentsComponent } from './previously-started-documents/previously-started-documents.component';
 import { PriorityQueueComponent } from './priority-queue/priority-queue.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
+import { GridsterModule } from 'angular-gridster2';
+import { StationWidgetComponent } from './widgets/station-widget/station-widget.component';
+import { MatCardModule } from '@angular/material/card';
+import { HeaderMenuComponent } from './dashboard-menu/header-menu/header-menu.component';
+import { OptionsMenuComponent } from './dashboard-menu/options-menu/options-menu.component';
+import { ExpansionMenuComponent } from './dashboard-menu/expansion-menu/expansion-menu.component';
+import { MenuComponent } from './dashboard-menu/menu/menu.component';
+import { LoadingIndicatorModule } from '../shared/loading-indicator/loading-indicator.module';
+import { RosterModule } from '../shared/roster/roster.module';
+import { StationDocumentsModalModule } from '../shared/station-documents-modal/station-documents-modal.module';
 
 @NgModule({
   declarations: [
@@ -21,13 +30,22 @@ import { MatInputModule } from '@angular/material/input';
     MyStationsComponent,
     PreviouslyStartedDocumentsComponent,
     PriorityQueueComponent,
+    StationWidgetComponent,
+    HeaderMenuComponent,
+    OptionsMenuComponent,
+    ExpansionMenuComponent,
+    MenuComponent,
   ],
   imports: [
     CommonModule,
     DashboardRoutingModule,
-    SharedModule,
     MatInputModule,
     ReactiveFormsModule,
+    GridsterModule,
+    MatCardModule,
+    LoadingIndicatorModule,
+    RosterModule,
+    StationDocumentsModalModule,
   ],
 })
 export class DashboardModule {}
