@@ -6,8 +6,17 @@ import { PriorityQueueComponent } from '../priority-queue/priority-queue.compone
 import { PreviouslyStartedDocumentsComponent } from '../previously-started-documents/previously-started-documents.component';
 import { MyStationsComponent } from '../my-stations/my-stations.component';
 import { StationService } from 'src/app/core/station.service';
-import { MockStationService, MockUserService } from 'src/mocks';
+import {
+  MockErrorService,
+  MockStationService,
+  MockUserService,
+  MockDashboardService,
+} from 'src/mocks';
 import { UserService } from 'src/app/core/user.service';
+import { ErrorService } from 'src/app/core/error.service';
+import { SplitService } from 'src/app/core/split.service';
+import { DashboardService } from '../dashboard.service';
+import { throwError } from 'rxjs';
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
@@ -25,6 +34,9 @@ describe('DashboardComponent', () => {
       providers: [
         { provide: StationService, useClass: MockStationService },
         { provide: UserService, useClass: MockUserService },
+        { provide: ErrorService, useClass: MockErrorService },
+        { provide: SplitService },
+        { provide: DashboardService, useClass: MockDashboardService },
       ],
     }).compileComponents();
   });
@@ -37,5 +49,35 @@ describe('DashboardComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should call service from return data widgets', () => {
+    const spyService = spyOn(
+      TestBed.inject(DashboardService),
+      'getDashboardWidgets'
+    ).and.callThrough();
+
+    component.ngOnInit();
+
+    expect(spyService).toHaveBeenCalled();
+  });
+
+  it('should catch error if petition to service fails', () => {
+    spyOn(
+      TestBed.inject(DashboardService),
+      'getDashboardWidgets'
+    ).and.returnValue(
+      throwError(() => {
+        throw new Error();
+      })
+    );
+
+    const spyError = spyOn(
+      TestBed.inject(ErrorService),
+      'displayError'
+    ).and.callThrough();
+
+    component.ngOnInit();
+    expect(spyError).toHaveBeenCalled();
   });
 });
