@@ -2,11 +2,11 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MockComponent } from 'ng-mocks';
 import { DocumentService } from 'src/app/core/document.service';
 import { ErrorService } from 'src/app/core/error.service';
-import { ConnectedStationPaneComponent } from 'src/app/detail/connected-station-pane/connected-station-pane.component';
-import { DocumentInfoHeaderComponent } from 'src/app/detail/document-info-header/document-info-header.component';
+import { ConnectedStationPaneComponent } from 'src/app/shared/connected-station-pane/connected-station-pane.component';
+import { DocumentInfoHeaderComponent } from 'src/app/shared/document-info-header/document-info-header.component';
 import { DocumentTemplateComponent } from 'src/app/document/document-template/document-template.component';
-import { StationInfoHeaderComponent } from 'src/app/detail/station-info-header/station-info-header.component';
-import { SubHeaderComponent } from 'src/app/detail/sub-header/sub-header.component';
+import { StationInfoHeaderComponent } from 'src/app/shared/station-info-header/station-info-header.component';
+import { SubHeaderComponent } from 'src/app/shared/sub-header/sub-header.component';
 import { DocumentComponent } from './document.component';
 import {
   MockDocumentService,
@@ -15,7 +15,7 @@ import {
 } from 'src/mocks';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { RouterTestingModule } from '@angular/router/testing';
-import { DetailDrawerComponent } from 'src/app/detail/detail-drawer/detail-drawer.component';
+import { DetailDrawerComponent } from 'src/app/shared/detail-drawer/detail-drawer.component';
 import { DashboardComponent } from 'src/app/dashboard/dashboard/dashboard.component';
 import { LoadingIndicatorComponent } from 'src/app/shared/loading-indicator/loading-indicator.component';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
@@ -23,11 +23,7 @@ import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { PopupService } from 'src/app/core/popup.service';
 import { Router } from '@angular/router';
-import {
-  DocumentAnswer,
-  DocumentAutoFlow,
-  QuestionFieldType,
-} from 'src/models';
+import { DocumentAutoFlow, QuestionFieldType } from 'src/models';
 import { forkJoin, of } from 'rxjs';
 import { MatExpansionModule } from '@angular/material/expansion';
 
@@ -398,6 +394,37 @@ describe('DocumentComponent', () => {
     expect(btnFlow.disabled).toBeFalsy();
   });
 
+  it('should called saveAnswers service when saving document changes', () => {
+    const expectedAnswer = component.documentAnswer;
+    const spySaveAnswerDocument = spyOn(
+      TestBed.inject(DocumentService),
+      'saveDocumentAnswer'
+    ).and.callThrough();
+    component.saveDocumentChanges();
+    expect(spySaveAnswerDocument).toHaveBeenCalledOnceWith(
+      component.documentInformation.documentRithmId,
+      expectedAnswer
+    );
+  });
+
+  it('should called UpdateDocumentName service when saving document changes', () => {
+    const documentService = TestBed.inject(DocumentService);
+    documentService.documentName$.next({
+      baseName: 'New Document Name',
+      appendedName: '',
+    });
+    const documentName = 'New Document Name';
+    const spyUpdateDocumentName = spyOn(
+      TestBed.inject(DocumentService),
+      'updateDocumentName'
+    ).and.callThrough();
+    component.saveDocumentChanges();
+    expect(spyUpdateDocumentName).toHaveBeenCalledOnceWith(
+      component.documentInformation.documentRithmId,
+      documentName
+    );
+  });
+
   it('should called service to save answers and auto flow the document', () => {
     const expectedAnswer = component.documentAnswer;
 
@@ -407,9 +434,20 @@ describe('DocumentComponent', () => {
       testMode: false,
     };
 
+    const documentService = TestBed.inject(DocumentService);
+    documentService.documentName$.next({
+      baseName: 'New Document Name',
+      appendedName: '',
+    });
+    const documentName = 'New Document Name';
+
     const spySaveAnswerDocument = spyOn(
       TestBed.inject(DocumentService),
       'saveDocumentAnswer'
+    ).and.callThrough();
+    const spyUpdateDocumentName = spyOn(
+      TestBed.inject(DocumentService),
+      'updateDocumentName'
     ).and.callThrough();
     const spySaveAutoFlowDocument = spyOn(
       TestBed.inject(DocumentService),
@@ -421,6 +459,10 @@ describe('DocumentComponent', () => {
     expect(spySaveAnswerDocument).toHaveBeenCalledOnceWith(
       component.documentInformation.documentRithmId,
       expectedAnswer
+    );
+    expect(spyUpdateDocumentName).toHaveBeenCalledOnceWith(
+      component.documentInformation.documentRithmId,
+      documentName
     );
     expect(spySaveAutoFlowDocument).toHaveBeenCalledOnceWith(expectAutoFlow);
   });
@@ -435,54 +477,6 @@ describe('DocumentComponent', () => {
 
     button.click();
 
-    expect(spyMethod).toHaveBeenCalled();
-  });
-
-  it('should test method to save document answer', () => {
-    const expectedAnswers: DocumentAnswer[] = [
-      {
-        questionRithmId: 'Dev 1',
-        documentRithmId: '123-654-789',
-        stationRithmId: '741-951-753',
-        value: 'Answer Dev',
-        file: 'dev.txt',
-        filename: 'dev',
-        type: QuestionFieldType.Email,
-        rithmId: '789-321-456',
-        questionUpdated: true,
-      },
-      {
-        questionRithmId: 'Dev 2',
-        documentRithmId: '123-654-789-856',
-        stationRithmId: '741-951-753-741',
-        value: 'Answer Dev2',
-        file: 'dev2.txt',
-        filename: 'dev2',
-        type: QuestionFieldType.City,
-        rithmId: '789-321-456-789',
-        questionUpdated: false,
-      },
-    ];
-    component.documentAnswer = expectedAnswers;
-
-    const spyQuestionAnswer = spyOn(
-      TestBed.inject(DocumentService),
-      'saveDocumentAnswer'
-    ).and.callThrough();
-    component.saveDocumentAnswer();
-    expect(spyQuestionAnswer).toHaveBeenCalledWith(
-      component.documentInformation.documentRithmId,
-      component.documentAnswer
-    );
-  });
-
-  it('should call the save method when the save button is clicked', () => {
-    component.documentLoading = false;
-    const spyMethod = spyOn(component, 'saveDocumentAnswer').and.callThrough();
-    fixture.detectChanges();
-    const buttonSave =
-      fixture.debugElement.nativeElement.querySelector('#document-save');
-    buttonSave.click();
     expect(spyMethod).toHaveBeenCalled();
   });
 
