@@ -5,12 +5,14 @@ import { ForgotPasswordComponent } from './forgot-password/forgot-password.compo
 import { AccountCreateComponent } from './account-create/account-create.component';
 import { PasswordResetComponent } from './password-reset/password-reset.component';
 import { EntryRoutingModule } from './entry-routing.module';
-import { SharedModule } from '../shared/shared.module';
 
 import { MatCardModule } from '@angular/material/card';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import { ReactiveFormsModule } from '@angular/forms';
+import { LoadingIndicatorModule } from '../shared/loading-indicator/loading-indicator.module';
+import { UserFormModule } from '../shared/user-form/user-form.module';
+import { PasswordRequirementsModule } from '../shared/password-requirements/password-requirements.module';
 
 @NgModule({
   declarations: [
@@ -22,11 +24,13 @@ import { ReactiveFormsModule } from '@angular/forms';
   imports: [
     CommonModule,
     EntryRoutingModule,
-    SharedModule,
     MatCardModule,
     MatInputModule,
     MatButtonModule,
     ReactiveFormsModule,
+    LoadingIndicatorModule,
+    UserFormModule,
+    PasswordRequirementsModule,
   ],
 })
 export class EntryModule {}
