@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Point } from 'src/models';
+import { CONNECTION_LINE_WIDTH } from './map-constants';
 import { MapService } from './map.service';
 
 /**
@@ -24,9 +25,26 @@ export class MapBoundaryService {
   drawBox(minBoundaryCoords: Point, maxBoundaryCoords: Point): void {
     this.canvasContext = this.mapService.canvasContext;
     if (!this.canvasContext) {
-      throw new Error(
-        'Cannot draw flow boundary line if context is not defined'
-      );
+      throw new Error('Cannot draw boundary line if context is not defined');
     }
+    const ctx = this.canvasContext;
+    const strokeColor = '#1b4387';
+
+    ctx.setLineDash([1, 5]);
+    ctx.beginPath();
+    ctx.lineCap = 'round';
+    ctx.strokeStyle = strokeColor;
+
+    ctx.moveTo(minBoundaryCoords.x, minBoundaryCoords.y);
+    ctx.lineTo(minBoundaryCoords.x, minBoundaryCoords.y);
+    ctx.lineTo(maxBoundaryCoords.x, minBoundaryCoords.y);
+    ctx.lineTo(maxBoundaryCoords.x, maxBoundaryCoords.y);
+    ctx.lineTo(minBoundaryCoords.x, maxBoundaryCoords.y);
+    ctx.lineTo(minBoundaryCoords.x, minBoundaryCoords.y);
+
+    ctx.lineWidth = CONNECTION_LINE_WIDTH;
+    ctx.closePath();
+    ctx.stroke();
+    ctx.setLineDash([]);
   }
 }
