@@ -6,7 +6,7 @@ import { UserService } from 'src/app/core/user.service';
 import { User, OrganizationInfo, MapMode } from 'src/models';
 import { MapService } from '../map.service';
 import { Subject } from 'rxjs';
-import { HttpErrorResponse } from '@angular/common/http';
+import { HttpErrorResponse, HttpStatusCode } from '@angular/common/http';
 /**
  * Component for managing the toolbar on the map.
  */
@@ -40,7 +40,7 @@ export class MapToolbarComponent implements OnInit, OnDestroy {
     return (
       this.mapMode === MapMode.Build ||
       this.mapMode === MapMode.StationAdd ||
-      this.mapMode === MapMode.FlowAdd
+      this.mapMode === MapMode.StationGroupAdd
     );
   }
 
@@ -54,12 +54,12 @@ export class MapToolbarComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Add flow mode active.
+   * Add station group mode active.
    *
    * @returns Boolean.
    */
-  get flowAddActive(): boolean {
-    return this.mapMode === MapMode.FlowAdd;
+  get stationGroupAddActive(): boolean {
+    return this.mapMode === MapMode.StationGroupAdd;
   }
 
   constructor(
@@ -83,16 +83,16 @@ export class MapToolbarComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Sets the map to add flow mode in preparation for a flow to be selected.
+   * Sets the map to add station group mode in preparation for a station group to be selected.
    */
-  addFlow(): void {
-    if (!this.flowAddActive) {
-      this.mapService.mapMode$.next(MapMode.FlowAdd);
+  addStationGroup(): void {
+    if (!this.stationGroupAddActive) {
+      this.mapService.mapMode$.next(MapMode.StationGroupAdd);
       this.mapService.matMenuStatus$.next(true);
     } else {
       this.mapService.mapMode$.next(MapMode.Build);
     }
-    // TODO: Implement add flow
+    // TODO: Implement add station group.
   }
 
   /**
@@ -138,7 +138,7 @@ export class MapToolbarComponent implements OnInit, OnDestroy {
             "Something went wrong on our end and we're looking into it. Please try again in a little while.";
           if (error instanceof HttpErrorResponse) {
             switch (error.status) {
-              case 401:
+              case HttpStatusCode.Unauthorized:
                 errorMessage =
                   'The user does not have rights to access the map.';
             }
