@@ -5,18 +5,18 @@ import { SettingsRoutingModule } from './settings-routing.module';
 import { NotificationSettingsComponent } from './notification-settings/notification-settings.component';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
-import { SharedModule } from '../shared/shared.module';
 import { ReactiveFormsModule } from '@angular/forms';
+import { LoadingIndicatorModule } from '../shared/loading-indicator/loading-indicator.module';
 
 @NgModule({
   declarations: [AccountSettingsComponent, NotificationSettingsComponent],
   imports: [
     CommonModule,
     SettingsRoutingModule,
-    SharedModule,
     MatCardModule,
     MatButtonModule,
     ReactiveFormsModule,
+    LoadingIndicatorModule,
   ],
 })
 export class SettingsModule {}
