@@ -1,9 +1,22 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FlowLogicComponent } from './flow-logic.component';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { ConnectedStationInfo } from 'src/models';
+import {
+  ConnectedStationInfo,
+  FlowLogicRule,
+  OperandType,
+  OperatorType,
+  RuleType,
+} from 'src/models';
 import { RuleModalComponent } from '../rule-modal/rule-modal.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatStepperModule } from '@angular/material/stepper';
+import { StationService } from 'src/app/core/station.service';
+import { MockErrorService, MockStationService } from 'src/mocks';
+import { ErrorService } from 'src/app/core/error.service';
+import { MatSelectModule } from '@angular/material/select';
+import { of, throwError } from 'rxjs';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 describe('FlowLogicComponent', () => {
   let component: FlowLogicComponent;
@@ -18,8 +31,18 @@ describe('FlowLogicComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [MatDialogModule, BrowserAnimationsModule],
+      imports: [
+        MatDialogModule,
+        BrowserAnimationsModule,
+        MatStepperModule,
+        MatSelectModule,
+        MatSnackBarModule,
+      ],
       declarations: [FlowLogicComponent, RuleModalComponent],
+      providers: [
+        { provide: StationService, useClass: MockStationService },
+        { provide: ErrorService, useClass: MockErrorService },
+      ],
     }).compileComponents();
   });
 
@@ -68,5 +91,104 @@ describe('FlowLogicComponent', () => {
       btnOpenModal.click();
       expect(spyFunc).toHaveBeenCalled();
     });
+  });
+
+  it('should call the method that returns the logical flow rules of a station.', () => {
+    component.rithmId = rithmId;
+    const getStationFlowLogicRuleSpy = spyOn(
+      TestBed.inject(StationService),
+      'getStationFlowLogicRule'
+    ).and.callThrough();
+    component.ngOnInit();
+    expect(getStationFlowLogicRuleSpy).toHaveBeenCalledWith(rithmId);
+  });
+
+  it('should show error message when request for logical flow rules of a station fails.', () => {
+    spyOn(
+      TestBed.inject(StationService),
+      'getStationFlowLogicRule'
+    ).and.returnValue(
+      throwError(() => {
+        throw new Error();
+      })
+    );
+    const displayErrorSpy = spyOn(
+      TestBed.inject(ErrorService),
+      'displayError'
+    ).and.callThrough();
+    component.ngOnInit();
+    expect(displayErrorSpy).toHaveBeenCalled();
+  });
+
+  it('should not show the display message when there are rules.', () => {
+    const expectStationFlowLogic: FlowLogicRule = {
+      stationRithmId: rithmId,
+      destinationStationRithmId: '73d47261-1932-4fcf-82bd-159eb1a7243f',
+      flowRules: [
+        {
+          ruleType: RuleType.Or,
+          equations: [
+            {
+              leftOperand: {
+                type: OperandType.Field,
+                value: 'birthday',
+              },
+              operatorType: OperatorType.Before,
+              rightOperand: {
+                type: OperandType.Date,
+                value: '5/27/1982',
+              },
+            },
+          ],
+        },
+      ],
+    };
+    spyOn(
+      TestBed.inject(StationService),
+      'getStationFlowLogicRule'
+    ).and.returnValue(of(expectStationFlowLogic));
+    component.ngOnInit();
+    fixture.detectChanges();
+    const messageNotRules =
+      fixture.debugElement.nativeElement.querySelector('#there-not-rules');
+    expect(messageNotRules).toBeFalsy();
+  });
+
+  it('should show the display message when there are not rules.', () => {
+    const expectStationFlowLogic: FlowLogicRule = {
+      stationRithmId: rithmId,
+      destinationStationRithmId: '73d47261-1932-4fcf-82bd-159eb1a7243f',
+      flowRules: [],
+    };
+    const messageNotRulesDefault =
+      fixture.debugElement.nativeElement.querySelector('#there-not-rules');
+    expect(messageNotRulesDefault).toBeTruthy();
+    spyOn(
+      TestBed.inject(StationService),
+      'getStationFlowLogicRule'
+    ).and.returnValue(of(expectStationFlowLogic));
+    component.ngOnInit();
+    fixture.detectChanges();
+    const messageNotRules =
+      fixture.debugElement.nativeElement.querySelector('#there-not-rules');
+    expect(messageNotRules).toBeTruthy();
+  });
+
+  it('should show error if petition rules fails', () => {
+    spyOn(
+      TestBed.inject(StationService),
+      'getStationFlowLogicRule'
+    ).and.returnValue(
+      throwError(() => {
+        throw new Error();
+      })
+    );
+    component.ngOnInit();
+    fixture.detectChanges();
+    const reviewError = fixture.debugElement.nativeElement.querySelector(
+      '#flow-logic-rules-error'
+    );
+    expect(component.flowRuleError).toBeTrue();
+    expect(reviewError).toBeTruthy();
   });
 });
