@@ -1,6 +1,12 @@
-import { Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
-
+import { Component, Inject, OnInit } from '@angular/core';
+import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { BreakpointObserver } from '@angular/cdk/layout';
+import { StepperOrientation } from '@angular/material/stepper';
+import { Observable } from 'rxjs';
+import { first, map } from 'rxjs/operators';
+import { StationService } from 'src/app/core/station.service';
+import { ErrorService } from 'src/app/core/error.service';
+import { Question } from 'src/models';
 /**
  * Reusable component for displaying the information to add a new rule.
  */
@@ -9,11 +15,66 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
   templateUrl: './rule-modal.component.html',
   styleUrls: ['./rule-modal.component.scss'],
 })
-export class RuleModalComponent {
+export class RuleModalComponent implements OnInit {
   /** Station Rithm id. */
   stationRithmId = '';
 
-  constructor(@Inject(MAT_DIALOG_DATA) public rithmId: string) {
+  /** Orientation for stepper. */
+  stepperOrientation$: Observable<StepperOrientation>;
+
+  /** The value of the first operand. */
+  firstOperand = '';
+
+  /** Get current and previous Questions for Stations. */
+  questionStation: Question[] = [];
+
+  /** The value of the operator. */
+  operator = '';
+
+  constructor(
+    public dialogRef: MatDialogRef<RuleModalComponent>,
+    @Inject(MAT_DIALOG_DATA) public rithmId: string,
+    breakpointObserver: BreakpointObserver,
+    private stationService: StationService,
+    private errorService: ErrorService
+  ) {
     this.stationRithmId = rithmId;
+    this.stepperOrientation$ = breakpointObserver
+      .observe('(min-width: 800px)')
+      .pipe(map(({ matches }) => (matches ? 'horizontal' : 'vertical')));
+  }
+
+  /**
+   * Life cycle init the component.
+   */
+  ngOnInit(): void {
+    this.getStationQuestions();
+  }
+
+  /**
+   * Close rule Modal.
+   */
+  closeModal(): void {
+    this.dialogRef.close();
+  }
+
+  /**
+   * Get current and previous questions.
+   */
+  getStationQuestions(): void {
+    this.stationService
+      .getStationQuestions(this.stationRithmId, true)
+      .pipe(first())
+      .subscribe({
+        next: (questions) => {
+          this.questionStation = questions;
+        },
+        error: (error: unknown) => {
+          this.errorService.displayError(
+            "Something went wrong on our end and we're looking into it. Please try again in a little while.",
+            error
+          );
+        },
+      });
   }
 }
