@@ -6,7 +6,7 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatCheckboxHarness } from '@angular/material/checkbox/testing';
 import { MockComponent } from 'ng-mocks';
 import { StationService } from 'src/app/core/station.service';
-import { TextFieldComponent } from 'src/app/detail/text-field/text-field.component';
+import { TextFieldComponent } from 'src/app/shared/fields/text-field/text-field.component';
 import { MockStationService } from 'src/mocks';
 import { QuestionFieldType } from 'src/models';
 import { StationFieldComponent } from './station-field.component';
@@ -103,7 +103,8 @@ describe('StationFieldComponent', () => {
     expect(component.field.isRequired).toBeTrue();
   });
 
-  describe('label field', () => {
+  // TODO: Add tests for this describe; this doesn't do anything
+  xdescribe('label field', () => {
     beforeEach(() => {
       component.field = {
         rithmId: '3j4k-3h2j-hj4j',
@@ -134,7 +135,21 @@ describe('StationFieldComponent', () => {
       fixture.detectChanges();
     });
 
-    it('should automatically add an option to the array', () => {
+    it("should automatically add an option to the array if there aren't answers", () => {
+      component.field.possibleAnswers = [];
+      const addOptionSpy = spyOn(component, 'addOption').and.callThrough();
+      expect(addOptionSpy).toHaveBeenCalled;
+      expect(component.options.length).toBeGreaterThan(0);
+    });
+
+    it('should automatically fill the options if there are possibleAnswers', () => {
+      component.field.possibleAnswers = [
+        {
+          default: false,
+          rithmId: '03BCE692-C347-484D-8EB3-3A2716F80BAF',
+          text: 'Light Armor',
+        },
+      ];
       expect(component.options.length).toBeGreaterThan(0);
     });
   });
