@@ -1,4 +1,4 @@
-import { FlowMapData, StationMapData } from '.';
+import { StationGroupMapData, StationMapData } from '.';
 
 /**
  * Contains all the data displayed on the map.
@@ -7,6 +7,6 @@ export interface MapData {
   /** List of all stations in the map. */
   stations: StationMapData[];
 
-  /** List of all flows in the map. */
-  flows: FlowMapData[];
+  /** List of all station groups in the map. */
+  stationGroups: StationGroupMapData[];
 }
